refacto: clean archi. added task features
